Cleanup: Removing all hibernate ORM
