+++ src/main/groovy/com/cellarhq/CellarHQModule.groovy
@@ -26,19 +26,9 @@ class CellarHQModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(EmailAccountDAO).in(Scopes.SINGLETON)
-        bind(OAuthAccountDAO).in(Scopes.SINGLETON)
-        bind(CellarDAO).in(Scopes.SINGLETON)
-        bind(DrinkDAO).in(Scopes.SINGLETON)
-        bind(DrinkService).in(Scopes.SINGLETON)
+        bind(AccountService).in(Scopes.SINGLETON)
+        bind(CellarService).in(Scopes.SINGLETON)
         bind(OrganizationService).in(Scopes.SINGLETON)
-        bind(StyleDAO).in(Scopes.SINGLETON)
-        bind(StyleService).in(Scopes.SINGLETON)
-        bind(GlasswareDAO).in(Scopes.SINGLETON)
-        bind(GlasswareService).in(Scopes.SINGLETON)
-        bind(DrinkCategoryDAO).in(Scopes.SINGLETON)
-        bind(DrinkCategoryService).in(Scopes.SINGLETON)
-        bind(PasswordChangeRequestDAO).in(Scopes.SINGLETON)
 
         if (System.getenv(ENV_DEPLOYMENT) == ENV_DEPLOYMENT_PRODUCTION) {
             // TODO: This should get put into the ratpack configuration file...
