+++ src/main/groovy/com/cellarhq/domain/DrinkCategory.groovy
@@ -1,38 +1,13 @@
 package com.cellarhq.domain
 
-import com.cellarhq.ratpack.hibernate.Entity
+import groovy.transform.CompileStatic
+import groovy.transform.InheritConstructors
 
-import javax.persistence.CascadeType
-import javax.persistence.Column
-import javax.persistence.FetchType
-import javax.persistence.GeneratedValue
-import javax.persistence.GenerationType
-import javax.persistence.Id
-import javax.persistence.OneToMany
+@CompileStatic
+@InheritConstructors
+class DrinkCategory extends com.cellarhq.generated.tables.pojos.Category {
 
-@Entity(name = 'category')
-class DrinkCategory extends AbstractEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id
-
-    @Column(length = 100, nullable = false)
-    String name
-
-    @Column(length = 2048, nullable = true)
-    String description
-
-    @Column(nullable = false)
-    boolean searchable = true
-
-    @Column(name = 'brewery_db_id', nullable = true)
-    String breweryDbId
-
-    @Column(name = 'brewery_db_last_updated', nullable = true)
-    Date breweryDbLastUpdated
-
-    @OneToMany(cascade = CascadeType.ALL, mappedBy = 'category', fetch = FetchType.LAZY)
-    Set<Style> styles = []
-
+    DrinkCategory() {
+        searchable = true
+    }
 }
