+++ src/main/groovy/com/cellarhq/endpoints/auth/ChangePasswordEndpoint.groovy
@@ -3,8 +3,8 @@ package com.cellarhq.endpoints.auth
 import static ratpack.handlebars.Template.handlebarsTemplate
 
 import com.cellarhq.Messages
-import com.cellarhq.domain.jooq.EmailAccount
-import com.cellarhq.services.JooqAccountService
+import com.cellarhq.domain.EmailAccount
+import com.cellarhq.services.AccountService
 import com.cellarhq.util.LogUtil
 import com.cellarhq.util.SessionUtil
 import com.google.inject.Inject
@@ -21,11 +21,11 @@ import javax.validation.ValidatorFactory
 @Slf4j
 class ChangePasswordEndpoint extends GroovyHandler {
 
-    private final JooqAccountService accountService
+    private final AccountService accountService
     private final ValidatorFactory validatorFactory
 
     @Inject
-    ChangePasswordEndpoint(JooqAccountService accountService, ValidatorFactory validatorFactory) {
+    ChangePasswordEndpoint(AccountService accountService, ValidatorFactory validatorFactory) {
         this.accountService = accountService
         this.validatorFactory = validatorFactory
     }
