+++ src/main/groovy/com/cellarhq/services/CellarService.groovy
@@ -1,39 +1,49 @@
 package com.cellarhq.services
 
-import com.cellarhq.dao.CellarDAO
 import com.cellarhq.domain.Cellar
+import com.cellarhq.generated.Tables
+import com.cellarhq.generated.tables.records.CellarRecord
 import com.google.inject.Inject
 import groovy.transform.CompileStatic
+import org.jooq.DSLContext
 import org.pac4j.oauth.profile.twitter.TwitterProfile
 
-@Deprecated
-@CompileStatic
-class CellarService {
+import javax.sql.DataSource
+import java.sql.Timestamp
+import java.time.LocalDateTime
 
-    private final CellarDAO cellarDAO
+@CompileStatic
+class CellarService extends BaseJooqService {
 
     @Inject
-    CellarService(CellarDAO cellarDAO) {
-        this.cellarDAO = cellarDAO
-    }
-
-    Cellar save(Cellar cellar) {
-        cellarDAO.save(cellar)
+    CellarService(DataSource dataSource) {
+        super(dataSource)
     }
 
     Cellar get(Long id) {
-        cellarDAO.find(id)
+        (Cellar) jooq { DSLContext create ->
+            create.select()
+                    .from(Tables.CELLAR)
+                    .where(Tables.CELLAR.ID.eq(id))
+                    .fetchOneInto(Cellar)
+        }
     }
 
-    Iterable<Cellar> all() {
-        cellarDAO.findAll()
+    Cellar save(Cellar cellar) {
+        jooq { DSLContext create ->
+            CellarRecord cellarRecord = create.newRecord(Tables.CELLAR, cellar)
+            if (cellar.id) {
+                create.executeUpdate(cellarRecord)
+            } else {
+                create.executeInsert(cellarRecord)
+            }
+            cellarRecord.into(Cellar)
+        }
     }
 
-    void updateLoginStats(Cellar cellar, TwitterProfile twitterProfile) {
-        cellar.lastLogin = new Date()
-//        cellar.lastLoginIp // TODO  Ratpack remote address support
-
-        if (cellar.updateFromNetwork) {
+    void updateLoginStats(Cellar cellar, TwitterProfile twitterProfile = null) {
+        cellar.lastLogin = Timestamp.valueOf(LocalDateTime.now())
+        if (cellar.updateFromNetwork && twitterProfile) {
             cellar.with {
                 displayName = twitterProfile.displayName
                 location = twitterProfile.location
@@ -41,8 +51,6 @@ class CellarService {
                 bio = twitterProfile.description
             }
         }
-
-        cellarDAO.save(cellar)
+        save(cellar)
     }
-
 }
